Change In Design
Changing to have 2 primary files, overlay.py and usergui.py.

usergui.py will handle the settings menu and other menus that shoudn't be running in the background.

overlay.py is being written to run in the background and handle the overlay and connection to the ESP32 IMU. I'm attemting to write it in a way to allow the possibility of other intergrations such as a phone. but it's still going to be a pita till I can add that x3

Current most of the code is placeholder, allowing me to start working on the overlay and handling logic. Hopefully tho it should work for a POC.